feat: added revenue unit tests

+++ src/domain/modules/revenues/use-case/implementation/delete-revenue.use-case.impl.ts
@@ -1,5 +1,6 @@
 import { DeleteRevenueUseCase } from '@domain/modules/revenues/use-case/delete-revenue.use-case';
 import { Id } from '@domain/value-objects/id.value-object';
+import { RevenueNotFoundError } from '@domain/modules/revenues/error/revenue-not-found.error';
 import { RevenueRepository } from '@domain/modules/revenues/repository/revenue.repository';
 
 export class DeleteRevenueUseCaseImpl extends DeleteRevenueUseCase {
@@ -8,6 +9,11 @@ export class DeleteRevenueUseCaseImpl extends DeleteRevenueUseCase {
   }
 
   public async execute(revenueId: Id): Promise<void> {
+    const revenueExists = await this.revenueRepository.getById(revenueId);
+    if (!revenueExists) {
+      throw new RevenueNotFoundError(revenueId);
+    }
+
     await this.revenueRepository.deleteById(revenueId);
   }
 }
